+++ packages/in_app_purchase/in_app_purchase_storekit/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.3.9
 
-* Converted `storefront()`, `transactions()`, `addPayment()`, `canMakePayment` to pigeon.
+* Converts `storefront()`, `transactions()`, `addPayment()`, `canMakePayment` to pigeon.
 * Updates minimum iOS version to 12.0 and minimum Flutter version to 3.16.6.
 
 ## 0.3.8+1